Null check corrections
Assuming that the user can supply nulls as input. This addresses parts of https://github.com/AngusJohnson/Clipper2/issues/802

One remaining issue is that there is a null check against a value returned from PathFromStr(), but that method always returns a non-null value, so the check would appear to be meaningless.