Minor tweaks (#801)
* Floating point comparison

* Clean-up

The variables are initialized in the constructor, and this using line seems to be unused.

* Update Clipper.RectClip.cs

Variable initialized in constructor.

* This looked to be unnecessary

* group is not used, remove unused using statement.

min_area also not used.
_solutionTree initialized in constructor.

* Null check corrections

Assuming that the user can supply nulls as input. This addresses parts of https://github.com/AngusJohnson/Clipper2/issues/802

One remaining issue is that there is a null check against a value returned from PathFromStr(), but that method always returns a non-null value, so the check would appear to be meaningless.

* Clean up null checks

These seem to be redundant due to outer checks.

* More redundant null checks

Also remove one unused variable (result)

* Remove redundant base constructors

Remove redundant .ToString()

* Readability change

The long while statement made this confusing to understand. Use braces to make the end of the statement and the action more apparent.